smart-study-assistant-simple/requirements.txt
typer
streamlit
openai
python-dotenv